requirements.txt
absl-py==0.1.11
aniso8601==1.3.0
astor==0.6.2
bleach==1.5.0
certifi==2017.4.17
chardet==3.0.3
click==6.7
cycler==0.10.0
Django==2.0.3
docopt==0.6.2
Flask==0.12.2
Flask-RESTful==0.3.6
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
idna==2.5
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mysql-connector-python==8.0.5
numpy==1.14.2
opencv-python==3.3.0.10
pandas==0.22.0
Pillow==3.0.0
pipreqs==0.4.9
ply==3.11
protobuf==3.5.2.post1
pygame==1.9.2a0
PyMySQL==0.7.11
pyparsing==2.2.0
PySide==1.2.2
pyswip==0.2.3
python-dateutil==2.7.0
pytz==2018.3
qimage2ndarray==1.6
requests==2.16.5
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
urllib3==1.21.1
Werkzeug==0.14.1
yarg==0.1.9
pydub==0.23.0
